Added deposit and withdrawal GUI, updated createAccount

+++ BankApp/src/Account.java
@@ -1,9 +1,8 @@
 
 public interface Account {
-
 	
-	public void deposit(int amount);
-	public void withdraw(int amount);
+	public void deposit(double amount);
+	public void withdraw(double amount);
 	public void moneyTransfer(Account obj, int amount);
 	public boolean loginAuthentication(String name, String password);
 	public String expiryDateCaluclator();
